Feat: Adding library mode. Refactor: files moved in order to differentiate between library components and svelte-kit services
