+++ src/lib/ShotstackEditTemplate/ShotstackEditTemplateService.ts
