+++ src/lib/ShotstackEditTemplate/constants.ts
