+++ src/lib/ShotstackEditTemplate/types.ts
