+++ src/lib/ShotstackEditTemplate/validate.ts
