+++ src/lib/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Form } from './Form/Form.svelte'
