+++ vite.config.ts
@@ -2,11 +2,14 @@ import { sveltekit } from '@sveltejs/kit/vite';
 import { svelte } from '@sveltejs/vite-plugin-svelte'
 import type { UserConfig } from 'vite';
 import { resolve } from 'path'
+import sveltePreprocess from 'svelte-preprocess'
 
 function configOptions(environment: boolean): UserConfig {
 	switch (environment) {
 		case true: return {
-			plugins: [svelte()],
+			plugins: [svelte({
+				preprocess: [sveltePreprocess({ postcss: true })],
+			})],
 			build: {
 				outDir: 'package',
 				lib: {
